% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boundary.R
\name{Boundary}
\alias{Boundary}
\title{Checks the display change implementation in Boundary experiments (Rayner, 1975)}
\usage{
Boundary(data_list = "", maxtrial = 999, boundary_loc = "BOUNDARY",
  start_flag = "DISPLAY CHANGE STARTED",
  end_flag = "DISPLAY CHANGE COMPLETED")
}
\arguments{
\item{data_list}{Input of data files to be processed. This can be specified in three ways:
\enumerate{
\item a directory that contains all the files (it will select all files ending with ".asc",
and order them by participant number, if present).
\item Directory to a txt file that contains all the .asc data file names inside:
e.g., data_list= "C:/My Data/data_list.txt".
In the .txt file, the directory for each .asc data file should appear on a separate row,
e.g.: C:/My Data/subject1.asc /n
C:/My Data/subject2.asc
\item A directory to a single .asc file: e.g., data_list= "C:/My Data/subject1.asc".
}}

\item{maxtrial}{Maximum number of experimental trials in the experiment (default= 9999)}

\item{boundary_loc}{The name of the message flag in the data file for the pixel location of the boundary (if present)}

\item{start_flag}{start flag for the beginning of the display change}

\item{end_flag}{end flag indicating the completion of the display change}
}
\description{
Checks the implementation of 1 boundary change during sentence reading. Checks the timing of the display change,
as well as whether the boundary was crossed in a forwards saccade.
}
\author{
Martin R. Vasilev
}
