% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibr.R
\name{Calibr}
\alias{Calibr}
\title{Extracts the calibration accuracy information from asc files (for data quality
monitoring, etc.)}
\usage{
Calibr(data_list, keep_time_diff = 0)
}
\arguments{
\item{data_list}{Input of data files to be processed. This can be specified in three ways:
\enumerate{
\item a directory that contains all the files (it will select all files ending with ".asc",
and order them by participant number, if present).
\item Directory to a txt file that contains all the .asc data file names inside:
e.g., data_list= "C:/My Data/data_list.txt".
In the .txt file, the directory for each .asc data file should appear on a separate row,
e.g.: C:/My Data/subject1.asc /n
C:/My Data/subject2.asc
\item A directory to a single .asc file: e.g., data_list= "C:/My Data/subject1.asc".
}}
}
\value{
A data frame containing the calibration accuracy data
}
\description{
Extracts the calibration accuracy information from asc files (for data quality
monitoring, etc.)
}
\author{
Martin R. Vasilev
}
