% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Clean up of pre-processed data}
\usage{
cleanData(raw_fix)
}
\arguments{
\item{raw_fix}{Raw fixations that were pre-processed with the EMreading package}
}
\description{
Clean up of pre-processed data
}
\author{
Martin R. Vasilev
}
