% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractMessages.R
\name{ExtractMessages}
\alias{ExtractMessages}
\title{Extract message timestamps data from .asc files}
\usage{
ExtractMessages(data_list = NULL, maxtrial = 9999,
  message_name = "MSG")
}
\arguments{
\item{data_list}{Input of data files to be processed. This can be specified in three ways:
\enumerate{
\item a directory that contains all the files (it will select all files ending with ".asc",
and order them by participant number, if present).
\item Directory to a txt file that contains all the .asc data file names inside:
e.g., data_list= "C:/My Data/data_list.txt".
In the .txt file, the directory for each .asc data file should appear on a separate row,
e.g.: C:/My Data/subject1.asc /n
C:/My Data/subject2.asc
\item A directory to a single .asc file: e.g., data_list= "C:/My Data/subject1.asc".
}}

\item{maxtrial}{Maximum number of trials in the experiment (default= 9999)}

\item{message_name}{Name of the message flag in the data which needs to be extracted. If only one message
needs to be extracted, provide it as a string (e.g., message_name= "DISPLAY CHANGE STARTED"). If there is more
than one message, provide it as a vector of string (e.g., message_name= c('DISPLAY CHANGE STARTED',
'DISPLAY CHANGE COMPLETED')).}
}
\description{
This function reads in data from .asc files and returns the timestamps for when the message occured.
Currently supports up to 3 message at a time.
}
\author{
Martin R. Vasilev
}
