% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractSamples.R
\name{ExtractSamples}
\alias{ExtractSamples}
\title{Extract raw eye samples data from .asc files}
\usage{
ExtractSamples(data_list = NULL, maxtrial = 9999)
}
\arguments{
\item{data_list}{Input of data files to be processed. This can be specified in three ways:
\enumerate{
\item a directory that contains all the files (it will select all files ending with ".asc",
and order them by participant number, if present).
\item Directory to a txt file that contains all the .asc data file names inside:
e.g., data_list= "C:/My Data/data_list.txt".
In the .txt file, the directory for each .asc data file should appear on a separate row,
e.g.: C:/My Data/subject1.asc /n
C:/My Data/subject2.asc
\item A directory to a single .asc file: e.g., data_list= "C:/My Data/subject1.asc".
}}

\item{maxtrial}{Maximum number of trials in the experiment (default= 9999)}
}
\description{
This function reads in data from .asc files and returns the raw eye samples in a
data frame format (time, x position, y position, pupil size)
}
\author{
Martin R. Vasilev
}
