% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeDoctor_PadLines.R
\name{EyeDoctor_PadLines}
\alias{EyeDoctor_PadLines}
\title{Pad lines for EyeDoctor processing}
\usage{
EyeDoctor_PadLines(data_dir = NULL, paddingSize = 5)
}
\arguments{
\item{data_dir}{A directory containing asc data files to be processed}

\item{paddingSize}{How many characters to pad to the left and to the right of the line margin.
#'}
}
\value{
Nothing; the padded asc files are directly saved to the input directory
}
\description{
This function pads text lines in the raw asc files to enable processing of fixations that are just to the left or to the
right text margin. Please note that if lines are not padded, EyeDoctor maps all fixations outside the text area as -1, which 
makes it impossible to map them to the line area on which they occured.
}
\author{
Martin R. Vasilev
}
