% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequency.R
\name{Frequency}
\alias{Frequency}
\title{Adds frequency information for words in the text.}
\usage{
Frequency(data, database = "SUBTLEX-UK")
}
\arguments{
\item{data}{Dataframe containing the raw fixation data that were extracted with the
EMreading package}

\item{database}{A string indicating which frequency database to use ("SUBTLEX-UK" or "SUBTLEX-US")}
}
\description{
Output includes: Same data frame, but with the added frequency information
}
\author{
Martin R. Vasilev
}
