% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI.R
\name{GUI}
\alias{GUI}
\title{Graphical user interface for data pre-processing using R shiny}
\usage{
GUI()
}
\description{
Graphical user interface for data pre-processing using R shiny
}
\author{
Martin R. Vasilev
}
