% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_fix.R
\name{num_fix}
\alias{num_fix}
\title{Calculation of number of fixations per trial using raw fixation data.}
\usage{
num_fix(data)
}
\arguments{
\item{data}{Dataframe containing the raw fixation data that were extracted with the
EMreading package}
}
\description{
Output includes: number of 1st-pass fixations, number of 2nd-pass fixations, and number of all fixations in trial.
Note: if you need number of fixations PER WORD, this is given by the 'wordMeasures' function.
}
\author{
Martin R. Vasilev
}
