% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseData.R
\name{ParseData}
\alias{ParseData}
\title{Pre-processing of single-line reading data}
\usage{
ParseData(dat, SL = TRUE)
}
\arguments{
\item{dat}{data frame containing the raw fixations extracted with the readingET package:}

\item{SL}{a logical indicating whether this is a single-line experiment (1= yes, 0= no)}
}
\description{
This function reads in data from .asc files and performs the pre-processing.
}
\author{
Martin R. Vasilev
}
