% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocFromDA1.R
\name{preprocFromDA1}
\alias{preprocFromDA1}
\title{Pre-processing of data using DA1 files}
\usage{
preprocFromDA1(data_dir = NULL, ResX = 1920, ResY = 1080,
  maxtrial = 999, tBlink = 50, padding = 0)
}
\arguments{
\item{data_dir}{Input of data files to be processed, as a directory that contains BOTH the ASC
and the DA1 files for all subjects}

\item{ResX}{X screen resolution in pixels}

\item{ResY}{Y screen resolution in pixels}

\item{maxtrial}{Maximum number of trials in the experiment}

\item{tBlink}{Time in milliseconds for detecting blinks before or after fixation.
If there is a blink x milliseconds before or after the fixation, it will be marked
as having a blink. The default is 50 ms.}
}
\value{
A data frame containing the data
}
\description{
This function reads in data from .asc files and DA1 files (from EyeDoctor) and merges them together.
}
\author{
Martin R. Vasilev
}
