% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLpreproc.R
\name{SLpreproc}
\alias{SLpreproc}
\title{Pre-processing of single-line reading data}
\usage{
SLpreproc(data_list = "preproc/files.txt", ResX = 1920, ResY = 1080,
  maxtrial = 120, tBlink = 50, plot = FALSE, keepLastFix = TRUE)
}
\arguments{
\item{data_list}{Input of data files to be processed. This can be specified in three ways:
1) a directory that contains all the files (it will select all files ending with ".asc",
and order them by participant number, if present).
2) Directory to a txt file that contains all the .asc data file names inside:
e.g., data_list= "C:/My Data/data_list.txt".
In the .txt file, the directory for each .asc data file should appear on a separate row,
e.g.: C:/My Data/subject1.asc /n
C:/My Data/subject2.asc
3) A directory to a single .asc file: e.g., data_list= "C:/My Data/subject1.asc".}

\item{ResX}{X screen resolution in pixels}

\item{ResY}{Y screen resolution in pixels}

\item{maxtrial}{Maximum number of trials in the experiment}

\item{tBlink}{Time in milliseconds for detecting blinks before or after fixation. 
If there is a blink x milliseconds before or after the fixation, it will be marked
as having a blink. The default is 50 ms.}

\item{plot}{A logical indicating whether to plot the raw and re-aligned fixations
as an image file. The default is TRUE. If set to FALSE, no images will be plotted.
Note that plotting images will generally take longer time to pre-process the data.
The images are saved in the current working directory in the folder "img".}
}
\value{
A data frame containing the data
}
\description{
This function reads in data from .asc files and performs the pre-processing.
}
\author{
Martin R. Vasilev
}
