% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialTime.R
\name{trialTime}
\alias{trialTime}
\title{Extracts trial times from raw .asc files}
\usage{
trialTime(
  data_list,
  maxtrial = 999,
  startFlag = "SYNCTIME",
  endFlag = "DISPLAY OFF"
)
}
\arguments{
\item{data_list}{Input of data files to be processed. This can be specified in three ways:
\enumerate{
\item a directory that contains all the files (it will select all files ending with ".asc",
and order them by participant number, if present).
\item Directory to a txt file that contains all the .asc data file names inside:
e.g., data_list= "C:/My Data/data_list.txt".
In the .txt file, the directory for each .asc data file should appear on a separate row,
e.g.: C:/My Data/subject1.asc /n
C:/My Data/subject2.asc
\item A directory to a single .asc file: e.g., data_list= "C:/My Data/subject1.asc".
}}

\item{maxtrial}{Maximum number of trials in the experiment}

\item{startFlag}{The flag in the data that indicates the start of trial (default: SYNCTIME in Eyetrack/Eyelink)}

\item{endFlag}{The flag in the data that indicates the end of the trial (default: DISPLAY OFF in Eyetrack)}
}
\description{
Extracts trial times from raw .asc files
}
\author{
Martin R. Vasilev
}
