% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordMeasures.R
\name{wordMeasures}
\alias{wordMeasures}
\title{Calculation of word-level fixation measures (FFD, SFD, GD, TVT) using raw fixation data.}
\usage{
wordMeasures(data)
}
\arguments{
\item{data}{Dataframe containing the raw fixation data that were extracted with the
EMreading package}
}
\description{
Calculation of word-level fixation measures (FFD, SFD, GD, TVT) using raw fixation data.
}
\author{
Martin R. Vasilev
}
