% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialTime.R
\name{trialTime}
\alias{trialTime}
\title{Extracts trial times from raw .asc files}
\usage{
trialTime(data_list, maxtrial = 999)
}
\arguments{
\item{data_list}{Input of data files to be processed. This can be specified in three ways:
1) a directory that contains all the files (it will select all files ending with ".asc",
and order them by participant number, if present).
2) Directory to a txt file that contains all the .asc data file names inside:
e.g., data_list= "C:/My Data/data_list.txt".
In the .txt file, the directory for each .asc data file should appear on a separate row,
e.g.: C:/My Data/subject1.asc /n
C:/My Data/subject2.asc
3) A directory to a single .asc file: e.g., data_list= "C:/My Data/subject1.asc".}

\item{maxtrial}{Maximum number of trials in the experiment}
}
\description{
Extracts trial times from raw .asc files
}
\author{
Martin R. Vasilev
}
